Add hook to get position of selection; better highlight of selection
