+++ packages/volto-slate/src/editor/less/editor.less
@@ -120,8 +120,8 @@
   }
 
   .highlight-selection {
-    background-color: grey;
-    color: white;
+    background-color: #8080803d;
+    // color: white;
   }
 
   .ui.input.editor-link input {
