+++ packages/volto-slate/src/hooks/index.js
@@ -1,2 +1,3 @@
 export { useIsomorphicLayoutEffect } from './useIsomorphicLayoutEffect';
 export { useEditorContext } from './useEditorContext';
+export { useSelectionPosition } from './useSelectionPosition';
